Removed TextDecoder
Here I removed the need for using TextDecoder for processing the
location name.
I had to implment my own sorting function, which might not sort 100%
correct, but it passed the tests.
Without the TextDecoder my implementation is much faster.

I also fixed a bug, where I forgot to process the end of the file.